Made game use 60 FPS
Had to do some haxx since the internal physics was scaled to 24 FPS. In
order to make the game speed be the same, we introduced the concept of
"dt" where dt = 1.0 is "as if" the game were running at 24 fps.

All internal position, velocities, and even image indexing is using
doubles now and they are scaled to `dt`.